requirements.txt
fastapi
uvicorn[standard]
langchain>=0.3.0
langchain-groq>=0.3.0
langchain-community>=0.3.0
langchain-huggingface
langchain-text-splitters
sentence-transformers
langchain-chroma
chromadb>0.5.0
pypdf
python-multipart
python-dotenv
docx2txt
python-pptx
easyocr
unstructured
numpy>=1.24.0
